frist commit
I 06/23/25 I am learning about commands that we will be using, and the shortcut we will need to know. I will need to create better habits to help better my success in college and in this program. I want to look more into javascript, so I need to get myself more disiplined so I can invest more time into learning it. This will be difficult but most good things are found on the road least tread.